+++ db/migrate/20190904004415_add_keywords_to_documents.rb
@@ -1,0 +1,5 @@
+class AddKeywordsToDocuments < ActiveRecord::Migration[6.0]
+  def change
+    add_column :documents, :keywords, :string
+  end
+end