+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_09_03_233503) do
+ActiveRecord::Schema.define(version: 2019_09_04_004415) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -61,6 +61,7 @@ ActiveRecord::Schema.define(version: 2019_09_03_233503) do
     t.boolean "featured", default: false, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.string "keywords"
     t.index ["byte_size"], name: "index_documents_on_byte_size"
     t.index ["content_type"], name: "index_documents_on_content_type"
     t.index ["download_count"], name: "index_documents_on_download_count"