+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **Document Changes**
   - Reduced size of folder names in documents sidebar
+  - Admins can now set keywords for documents to improve search results for
+    documents
 - **User Changes**
   - Admins can now associate a Staff ID with a user that will display in the
     directory and in the menu bar